Separate out template functions and decorators
Move template functions and decorators to template_functions.py and
decorators.py, to help keep app.py focused on routing.
